users and people can comment.
As I'm getting more into this, I'm not sure if I should distinguish between the
owner of the account (User), and the people in the system in this way.

An Account model that holds the details of the account, with many users in it
might be better.  Users of course have varying permissions and levels, and it
would allow for a more even interface.

Not sure though.